add Sandbox gate to fleet UI
this adds the SandboxGate component to org settings, user settings, and user management features
