+++ frontend/pages/admin/AppSettingsPage/_styles.scss
@@ -6,307 +6,4 @@
     color: $core-fleet-black;
     @include sticky-settings-description;
   }
-
-  h2 {
-    font-size: $medium;
-    font-weight: $regular;
-    color: $core-fleet-black;
-    border-bottom: solid 1px $ui-fleet-blue-15;
-    margin: 0 0 $pad-xxlarge;
-  }
-
-  small {
-    font-size: $x-small;
-    font-weight: $regular;
-    color: $core-fleet-black;
-  }
-
-  &__settings-form {
-    display: flex;
-  }
-
-  &__form-nav-list {
-    position: -webkit-sticky;
-    position: sticky;
-    // this is the spacing needed to make the sticky form nav position correctly when scrolling
-    // TODO: find a way to calculate these sticky positions this and use variables.
-    //  will be tedious to update otherwise.
-    top: 217px;
-    width: 178px;
-    margin: 0;
-    padding: 0 110px 0 0;
-    list-style: none;
-    font-size: $x-small;
-
-    li {
-      margin-bottom: $pad-medium;
-    }
-
-    a {
-      color: $core-fleet-black;
-      font-weight: $regular;
-      text-decoration: none;
-      cursor: pointer;
-
-      &:hover {
-        color: $core-vibrant-blue;
-      }
-    }
-
-    .active-nav {
-      font-weight: $bold;
-    }
-  }
-
-  .form-field__label {
-    .buttons {
-      top: 22px;
-    }
-  }
-}
-
-.app-config-form {
-  width: 100%;
-
-  &__config-docs {
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding: $pad-medium;
-    border-radius: $border-radius;
-    border: 1px solid #d9d9fe;
-    background-color: $ui-vibrant-blue-10;
-    margin-bottom: $pad-medium;
-    font-size: $x-small;
-
-    p {
-      margin: 0;
-      font-size: $x-small;
-    }
-  }
-
-  &__learn-more {
-    color: $core-vibrant-blue;
-    font-size: $x-small;
-    font-weight: $bold;
-    text-decoration: none;
-
-    &--inline {
-      margin-left: $pad-medium;
-    }
-    img {
-      width: 12px;
-      height: 12px;
-      margin-left: $pad-small;
-    }
-  }
-
-  .form-field {
-    width: 100%;
-
-    &__label {
-      font-weight: $bold;
-      margin-bottom: $pad-xsmall;
-      width: 100%;
-    }
-
-    &__hint {
-      font-size: $x-small;
-      font-style: italic;
-    }
-  }
-
-  &__section {
-    @include clearfix;
-    margin: 0 0 $pad-large;
-
-    .upcaret::after {
-      content: url("../assets/images/icon-collapse-black-16x16@2x.png");
-      transform: scale(0.5);
-      border-radius: 0px;
-      position: relative;
-      top: 8px;
-      margin: $pad-xsmall;
-    }
-
-    .downcaret::after {
-      content: url("../assets/images/icon-chevron-black-16x16@2x.png");
-      transform: scale(0.5);
-      border-radius: 0px;
-      position: relative;
-      top: 8px;
-      margin: $pad-xsmall;
-      margin-right: 18px;
-    }
-
-    h2 {
-      padding-bottom: $pad-small;
-      max-width: 65%;
-    }
-
-    .smtp-options {
-      font-size: 15px;
-      font-weight: $bold;
-      color: $core-fleet-black;
-      padding-left: 15px;
-
-      em {
-        font-style: normal;
-      }
-
-      &--configured {
-        em {
-          color: $ui-success;
-        }
-      }
-
-      &--notconfigured {
-        em {
-          color: $ui-error;
-        }
-      }
-    }
-  }
-
-  &__section-description {
-    font-size: $x-small;
-    color: $core-fleet-black;
-    width: 60%;
-  }
-
-  &__yaml {
-    .app-config-form__section-description {
-      width: initial;
-    }
-
-    h2 {
-      max-width: initial;
-    }
-  }
-
-  &__inputs {
-    width: 60%;
-    float: left;
-    padding-right: $pad-small;
-    box-sizing: border-box;
-
-    .input-field {
-      width: 100%;
-    }
-
-    &--smtp {
-      margin: 0 0 $pad-medium;
-
-      .form-field {
-        width: 29%;
-        float: right;
-
-        &:first-child {
-          float: left;
-          width: 69%;
-        }
-
-        &:nth-child(2) {
-          padding-top: 10px;
-        }
-
-        &--checkbox {
-          clear: both;
-          width: 100%;
-          float: none;
-        }
-      }
-    }
-    &--usage {
-      margin-top: $pad-large;
-    }
-    &--preview {
-      margin: $pad-medium 0;
-    }
-    &--webhook {
-      margin-bottom: $pad-large;
-    }
-  }
-
-  &__details {
-    float: right;
-    width: 40%;
-    height: 87px;
-
-    .icon-tooltip {
-      margin: $pad-xlarge 0;
-    }
-
-    .hint {
-      color: $core-fleet-black;
-
-      &--brand {
-        color: $core-vibrant-blue;
-      }
-    }
-  }
-
-  &__avatar-preview {
-    text-align: center;
-
-    img {
-      border-radius: 20%;
-      height: 120px;
-      width: 120px;
-      border: 1px solid $ui-fleet-blue-15;
-      background-color: $ui-light-grey;
-      position: relative;
-      bottom: -20px;
-    }
-
-    p {
-      color: $core-fleet-purple;
-      font-size: 18px;
-      font-weight: $bold;
-      margin-top: 0;
-    }
-  }
-
-  &__smtp-section {
-    @include clearfix;
-  }
-
-  &__component-label {
-    margin: 24px 0 0;
-    font-size: $x-small;
-  }
-
-  &__yaml {
-    width: calc(90% - 50px);
-    margin-bottom: $pad-medium;
-  }
-
-  &__usage-stats-preview-modal,
-  &__host-status-webhook-preview-modal {
-    .flex-end {
-      margin-top: 20px;
-      display: flex;
-      justify-content: flex-end;
-    }
-  }
-
-  &__advanced-options {
-    margin-bottom: $pad-large;
-  }
-
-  &__transparency {
-    color: $core-vibrant-blue;
-    font-size: $x-small;
-    font-weight: $bold;
-    text-decoration: none;
-
-    img {
-      width: 12px;
-      height: 12px;
-    }
-  }
-
-  .component__tooltip-wrapper {
-    margin-bottom: $pad-xsmall;
-  }
 }
