+++ frontend/pages/admin/AppSettingsPage/components/OrgSettingsForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./OrgSettingsForm";
