+++ frontend/pages/admin/UserManagementPage/_styles.scss
@@ -4,126 +4,8 @@
     font-size: $x-small;
     color: $core-fleet-black;
     @include sticky-settings-description;
-    padding-bottom: $pad-medium;
   }
 
-  &__wrapper {
-    border: solid 1px $ui-fleet-blue-15;
-    border-radius: 6px;
-    margin-top: $pad-medium;
-  }
-
-  &__header {
-    font-size: $large;
-  }
-
-  &__heading-wrapper {
-    width: 100%;
-  }
-
-  &__smtp-warning-wrapper {
-    width: 100%;
-  }
-
-  &__config-button {
-    text-decoration: none;
-    color: $core-vibrant-blue;
-    font-size: $x-small;
-    margin-left: $pad-medium;
-  }
-
-  &__user-count {
-    font-size: $x-small;
-    font-weight: $bold;
-    color: $core-fleet-black;
-    margin: 0;
-  }
-
-  &__add-user-wrap {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-  }
-
-  &__users {
-    display: flex;
-    flex-wrap: wrap;
-    clear: both;
-  }
-
-  &__invite-modal {
-    &.modal__modal_container {
-      width: 540px;
-    }
-  }
-  .data-table-block {
-    .data-table__table {
-      thead {
-        .name__header {
-          width: $col-md;
-        }
-        .role__header {
-          width: $col-md;
-        }
-        .teams__header {
-          display: none;
-          width: 0;
-        }
-        .status__header {
-          display: none;
-          width: 0;
-        }
-        .actions__header {
-          width: auto;
-        }
-        @media (min-width: $break-990) {
-          .teams__header {
-            display: table-cell;
-            width: $col-md;
-          }
-        }
-        @media (min-width: $break-1400) {
-          .status__header {
-            display: table-cell;
-            width: $col-md;
-          }
-        }
-      }
-      tbody {
-        .name__cell {
-          max-width: $col-md;
-        }
-        .role__cell {
-          max-width: $col-md;
-        }
-        .teams__cell {
-          display: none;
-          max-width: $col-md;
-          .grey-cell {
-            color: $ui-fleet-black-50;
-            font-style: italic;
-          }
-        }
-        .status__cell {
-          display: none;
-          max-width: $col-md;
-        }
-        .actions__cell {
-          width: auto;
-        }
-        @media (min-width: $break-990) {
-          .teams__cell {
-            display: table-cell;
-          }
-        }
-        @media (min-width: $break-1400) {
-          .status__cell {
-            display: table-cell;
-          }
-        }
-      }
-    }
-  }
   .Select-menu-outer {
     position: absolute;
     left: -50px !important; // Using !important to override react select styling
