+++ frontend/pages/admin/UserManagementPage/components/UsersTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./UsersTable";
